package/gcc-bare-metal: Match ABI to selected architecture
The default abi and arch for gcc-bare-metal may not match the selected
architecture in buildroot.

Copy the fragments from buildroot/package/gcc/gcc.mk that set the
ABI, ARCH, FPU, FLOAT_ABI, and SOFT_FLOAT variables from GCC_TARGET_*
to the gcc-bare-metal package to ensure that the correct values are
used when building the toolchain.

Signed-off-by: Jesse Taube <Mr.Bossman075@gmail.com>